use max migration for schema version
Rather than requiring the user to create and update the schema version
table in their schema.sql, that table will be automatically handled by
dbami when loading the schema similar to how it is for migrations. The
schema version value is now populated via the max migration version,
and means we absolutely require a starting migration now, so the 0
migration is created by `init`.

These changes also clean up handling of a rollback attempt of the base
migration by explicitly highlighting that it isn't supported.